% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdMethods.R
\name{as.rdClass}
\alias{as.rdClass}
\title{Coerce an object of class 'fdClass', or 'imgClass' to class 'rdClass'}
\usage{
as.rdClass(x, Sparsity_parameter = NULL)
}
\arguments{
\item{x}{An object of class 'fdClass', or 'imgClass'.}

\item{Sparsity_parameter}{Optional sparsity parameter to override the original.}
}
\value{
An object of class 'rdClass' with smoothing and grid attributes removed,
and sparsity parameter preserved or overridden.
}
\description{
Coerce an object of class 'fdClass', or 'imgClass' to class 'rdClass'
}
\examples{
img_object <- imgClass(image = list(matrix(rnorm(100), nr= 50),
                                    matrix(rnorm(100),nr = 50)),
                       argval = NULL,
                       Smoothing_parameter = NULL,
                       Sparsity_parameter = 0)

newrd <- as.rdClass(img_object, Sparsity_parameter = 1:10)
attr(newrd , "Sparsity_parameter")

}
