% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdMethods.R
\name{$.fdClass}
\alias{$.fdClass}
\title{Custom \code{$} operator for fdClass
Allows access to the underlying data matrix via \code{fd$matrix}}
\usage{
\method{$}{fdClass}(x, name)
}
\arguments{
\item{x}{An object of class 'fdClass'}

\item{name}{The name of the element to extract}
}
\description{
Custom \code{$} operator for fdClass
Allows access to the underlying data matrix via \code{fd$matrix}
}
