% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{$.hdClass}
\alias{$.hdClass}
\title{Custom \code{$} operator for hdClass
Returns a clean data.frame when using hd_obj$matrix}
\usage{
\method{$}{hdClass}(x, name)
}
\arguments{
\item{x}{An object of class 'hdClass'}

\item{name}{The field to extract}
}
\description{
Custom \code{$} operator for hdClass
Returns a clean data.frame when using hd_obj$matrix
}
