% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgMethod.R
\name{$.imgClass}
\alias{$.imgClass}
\title{Custom \code{$} operator for imgClass
Allows access to the underlying data matrix via \code{img$matrix}}
\usage{
\method{$}{imgClass}(x, name)
}
\arguments{
\item{x}{An object of class 'imgClass'}

\item{name}{The name of the element to extract}
}
\description{
Custom \code{$} operator for imgClass
Allows access to the underlying data matrix via \code{img$matrix}
}
