% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdMethods.R
\name{$.rdClass}
\alias{$.rdClass}
\title{Custom \code{$} operator for rdClass
Allows access to the underlying data matrix via \code{rd$matrix}}
\usage{
\method{$}{rdClass}(x, name)
}
\arguments{
\item{x}{An object of class 'rdClass'}

\item{name}{The name of the element to extract}
}
\description{
Custom \code{$} operator for rdClass
Allows access to the underlying data matrix via \code{rd$matrix}
}
