% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.fdClass}
\alias{is.fdClass}
\title{Check if an object is of class 'fdClass'}
\usage{
is.fdClass(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical; TRUE if the object inherits from class 'fdClass', FALSE otherwise.
}
\description{
Check if an object is of class 'fdClass'
}
