% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.hdClass}
\alias{is.hdClass}
\title{Check if an object is of class 'hdClass'}
\usage{
is.hdClass(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical; TRUE if the object inherits from class 'hdClass', FALSE otherwise.
}
\description{
Check if an object is of class 'hdClass'
}
