% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.imgClass}
\alias{is.imgClass}
\title{Check if an object is of class 'imgClass'}
\usage{
is.imgClass(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical; TRUE if the object inherits from class 'imgClass', FALSE otherwise.
}
\description{
Check if an object is of class 'imgClass'
}
