% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.rdClass}
\alias{is.rdClass}
\title{Check if an object is of class 'rdClass'}
\usage{
is.rdClass(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical; TRUE if the object inherits from class 'rdClass', FALSE otherwise.
}
\description{
Check if an object is of class 'rdClass'
}
