% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdMethods.R
\name{plot.rdClass}
\alias{plot.rdClass}
\title{Plot Method for rdClass Objects}
\usage{
\method{plot}{rdClass}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{rdClass}.}
}
\value{
No return value. This function is called for its side effect (plot).
}
\description{
Produces a scatter plot of regular data stored in an \code{rdClass} object.
Each column is plotted as a sequence of solid points.
}
\details{
This method visualizes the regular (non-functional) data in the \code{rdClass} object.
It shows the values in each column as solid dots, which is useful for examining patterns across observations or variables.
}
\examples{
rd_obj <- rdClass(matrix(rnorm(100), nrow = 10, ncol = 10))
plot(rd_obj)

}
