% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.R
\name{read.data}
\alias{read.data}
\title{Read ReMPCA Example Data}
\usage{
read.data(data = c("bike_day", "bike_hour"))
}
\arguments{
\item{data}{Character string specifying the dataset to load.
Available options are \code{"bike_day"} and \code{"bike_hour"}.}
}
\value{
The requested dataset as a data frame.
}
\description{
Loads one of the datasets included with the ReMPCA package.
}
\examples{
bike_day <- read.data("bike_day")
bike_hour <- read.data("bike_hour")

}
