% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{setSmoothnessParameter}
\alias{setSmoothnessParameter}
\title{Set Smoothness Tuning Parameter}
\usage{
setSmoothnessParameter(obj, Smoothing_parameter)
}
\arguments{
\item{obj}{An object of class \code{rdClass}, \code{fdClass}, \code{imgClass}, or \code{hdClass}.}

\item{Smoothing_parameter}{A numeric value or vector representing the smoothing parameter(s)
to assign to the object.}
}
\value{
The modified object with updated \code{Smoothing_parameter} attribute.
}
\description{
Assigns a smoothing tuning parameter to an object of class \code{rdClass}, \code{fdClass}, \code{imgClass}, or \code{hdClass}.
}
\details{
This function updates the \code{"Smoothing_parameter"} attribute of the given object.
Since R uses copy-on-modify semantics for S3 objects, users must reassign the object after calling this function:
\preformatted{
  obj <- setSmoothnessParameter(obj, c(0.01, 0.1, 1))
}
Without reassignment, the original object remains unchanged.
}
\examples{
fd_obj <- fdClass(matrix(rnorm(100), 10, 10))
fd_obj <- setSmoothnessParameter(fd_obj, c(0.01, 0.1, 1))
attr(fd_obj, "Smoothing_parameter")

}
