% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{setSparsityParameter}
\alias{setSparsityParameter}
\title{Set Sparsity Tuning Parameter}
\usage{
setSparsityParameter(obj, Sparsity_parameter)
}
\arguments{
\item{obj}{An object of class \code{rdClass}, \code{fdClass}, \code{imgClass}, or \code{hdClass}.}

\item{Sparsity_parameter}{A numeric vector of non-negative integers indicating sparsity tuning levels.
If \code{NULL}, a default sequence will be generated based on the object's dimensions.}
}
\value{
The modified object with updated sparsity parameters.
}
\description{
Assigns a sparsity tuning parameter to an object of class \code{rdClass}, \code{fdClass}, \code{imgClass}, or \code{hdClass}.
}
\details{
This function returns a modified version of the input object with an updated \code{"Sparsity_parameter"} attribute.
Due to R's copy-on-modify semantics for S3 objects, the user must reassign the object:
\preformatted{
  obj <- setSparsityParameter(obj, c(0, 2, 4))
}
The object will not be updated in-place unless reassigned.
}
\examples{
fd_obj <- fdClass(matrix(rnorm(100), 10, 10))
fd_obj <- setSparsityParameter(fd_obj, 0:5)
attr(fd_obj, "Sparsity_parameter")

}
