% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMFPCfunc.R
\name{SSMFPCfunc}
\alias{SSMFPCfunc}
\title{SSMFPCfunc Smooth and Sparse Multivariate Functional Principal Component Analysis}
\usage{
SSMFPCfunc(
  mvfd_obj,
  argval = NULL,
  centerfns = TRUE,
  num_pcs = 1,
  smooth_tuning = NULL,
  sparse_tuning_type = "soft",
  sparse_tuning = 0
)
}
\arguments{
\item{mvfd_obj}{A list of data matrices, where each one is considered of as a variable and observations are stored in the rows, and grid points are in the columns. It is also possible for the timeline or grid points to include the column name.}

\item{argval}{A list of grid points corresponding to each variable, where the length of each component matches the number of columns in the related data matrix.}

\item{centerfns}{A logical}

\item{num_pcs}{Logical: if True, it demeans the data before calculating the principal components.}

\item{smooth_tuning}{A data set of different combinations of alphas (smoothness parameters) for all variables.}

\item{sparse_tuning_type}{A character string specifying the sparse calculation method. Must be one of "soft" (default), "hard", or "SCAD".}

\item{sparse_tuning}{A number that shows the level of sparsity.}
}
\value{
PC scores, PC functions, ...
}
\description{
SSMFPCfunc Smooth and Sparse Multivariate Functional Principal Component Analysis
}
