% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{[.hdClass}
\alias{[.hdClass}
\title{Indexing Operator for hdClass}
\usage{
\method{[}{hdClass}(x, k = NULL)
}
\arguments{
\item{x}{An object of class \code{hdClass}.}

\item{k}{Variables indices. If \code{NULL}, all variables are included.}
}
\value{
A new \code{hdClass} object with subsetted data and inherited attributes.
}
\description{
Enables subsetting of an \code{hdClass} object by variables.
}
