% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgMethod.R
\name{[.imgClass}
\alias{[.imgClass}
\title{Indexing operator for imgClass}
\usage{
\method{[}{imgClass}(x, i = NULL)
}
\arguments{
\item{x}{An object of class \code{imgClass}.}

\item{i}{Row indices (Images). If \code{NULL}, all images are included.}
}
\value{
A new \code{imgClass} object with subsetted data and inherited attributes.
}
\description{
Enables subsetting of an \code{imgClass} object by rows.
}
