% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{*.hdClass}
\alias{*.hdClass}
\title{Multiply a \code{hdClass} Object by a Scalar}
\usage{
\method{*}{hdClass}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric value or a \code{hdClass} object.}

\item{e2}{A \code{hdClass} object or a scalar numeric value.}
}
\value{
A new \code{hdClass} object with elements scaled by the scalar value, and original attributes retained.
}
\description{
Performs element-wise multiplication between a scalar and a \code{hdClass} object.
All attributes and class information are preserved.
}
\examples{
obj <- hdClass(list(fdClass(matrix(1:10, ncol = 2))))
2 * obj

}
