% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgMethod.R
\name{*.imgClass}
\alias{*.imgClass}
\title{Multiply a \code{imgClass} Object by a Scalar}
\usage{
\method{*}{imgClass}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric value or an \code{imgClass} object.}

\item{e2}{An \code{imgClass} object or a scalar numeric value.}
}
\value{
A new \code{imgClass} object with each image scaled by the scalar value.
}
\description{
Multiplies each matrix (image) in an \code{imgClass} object by a scalar.
All attributes and class structure are retained.
}
\examples{
img <- imgClass(image = list(matrix(1:9, 3, 3)))
img_scaled <- 2 * img

}
