% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdMethods.R
\name{*.rdClass}
\alias{*.rdClass}
\title{Multiply a \code{rdClass} Object by a Scalar}
\usage{
\method{*}{rdClass}(e1, e2)
}
\arguments{
\item{e1}{A scalar numeric value or a \code{rdClass} object.}

\item{e2}{A \code{rdClass} object or a scalar numeric value.}
}
\value{
A \code{rdClass} object scaled by the numeric scalar.
}
\description{
Performs element-wise multiplication between a scalar and a \code{rdClass} object.
Attributes like sparsity settings are preserved.
}
\examples{
rd <- rdClass(matrix(1:12, nrow = 4))
rd_scaled <- rd * 0.5

}
