% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdMethods.R
\name{as.fdClass}
\alias{as.fdClass}
\title{Coerce an object of class 'rdClass', or 'imgClass' to class 'fdClass'.}
\usage{
as.fdClass(
  x,
  Sparsity_parameter = NULL,
  Smoothing_parameter = NULL,
  argval = NULL
)
}
\arguments{
\item{x}{An object of class 'rdClass', or 'imgClass'.}

\item{Sparsity_parameter}{Optional sparsity parameter to assign. If \code{NULL}, defaults are used.}

\item{Smoothing_parameter}{Optional smoothing parameter to assign. If \code{NULL}, defaults are used.}

\item{argval}{Optional vector of grid points for columns. If \code{NULL}, defaults are used.}
}
\value{
An object of class 'fdClass' with user-specified or inherited regularization parameters.
}
\description{
Coerce an object of class 'rdClass', or 'imgClass' to class 'fdClass'.
}
\examples{
img_object <- imgClass(image = list(matrix(rnorm(100), nr= 50),
                                    matrix(rnorm(100),nr = 50)),
                       argval = NULL,
                       Smoothing_parameter = NULL,
                       Sparsity_parameter = 0)

newfd <- as.fdClass(img_object)
newfd

}
