% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdClass.R
\name{fdClass}
\alias{fdClass}
\title{Functional Data Class}
\usage{
fdClass(data, argval = NULL, Smoothing_parameter = 0, Sparsity_parameter = 0)
}
\arguments{
\item{data}{A matrix representing the data, with rows indicating observations and columns representing grid points.}

\item{argval}{A vector of grid points, it assigns grid points to the columns, with a length equal to the number of columns in the data.
If \code{NULL}, grid points are automatically assigned from 0 to 1.}

\item{Smoothing_parameter}{: Smoothing parameter for columns. It can be:
\itemize{
\item A fixed number representing the smoothing parameter.
\item A vector of numerical values, which will undergo generalized cross-validation (GCV) to determine the optimal value.
\item Set to 0 for no smoothing.
\item If \code{NULL}, it analyzes a sequence of \code{2^seq(-30, 5, length.out = 10)} and attempts to tune it.
}}

\item{Sparsity_parameter}{\itemize{
\item A fixed number representing the level of sparsity for columns, or a vector of numerical values that will undergo cross-validation (CV) to determine the optimal value.
\item For no sparsity, set it to 0.
\item If \code{NULL}, the sparsity parameter will be tuned automatically.
}}
}
\description{
The \code{fdClass} class represents an element of functional data.
\itemize{
\item Data objects are constructed using matrices, where columns represent grid points (user-defined or NULL) and rows represent observations.
\item Users can assign both smoothing and sparsity parameters.
\item If the smoothing parameter is set to zero, no smoothing is applied. Otherwise,
users can specify a fixed smoothing value or provide a vector of values, which will be optimized using generalized cross-validation (GCV).
}
}
\examples{
# Example for Functional Data (fd)
fd_data <- matrix(rnorm(100), nrow = 10, ncol = 10)  # 10 rows, 10 columns
fd_object <- fdClass(data = fd_data,
                     argval = seq(0, 1, length.out = 10),  # Grid points for columns
                     Smoothing_parameter = 0.5,  # Custom smoothing parameter
                     Sparsity_parameter = 2)  # Custom sparsity parameter

# Display the created fd object
print(fd_object)
print(attr(fd_object, "GridPoints_v"))  # Display grid points for columns
print(attr(fd_object, "Smoothing_parameter"))  # Display smoothing parameter

}
