% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdMethods.R
\name{plot.fdClass}
\alias{plot.fdClass}
\title{Plot Method for fdClass Objects}
\usage{
\method{plot}{fdClass}(x)
}
\arguments{
\item{x}{An object of class \code{fdClass}.}
}
\value{
No return value. This function is called for its side effect (plot).
}
\description{
Generates a line plot of the functional data stored in an object of class \code{fdClass}.
}
\details{
This function uses \code{\link{matplot}} to visualize each observation (row) as a separate curve.
It provides a quick overview of the functional data structure stored in the \code{fdClass} object.
}
\examples{
fd_obj <- fdClass(matrix(sin(1:100 / 10), nrow = 10, ncol = 10))
plot(fd_obj)

}
