% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{plot.hdClass}
\alias{plot.hdClass}
\title{Plot Method for hdClass Objects}
\usage{
\method{plot}{hdClass}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hdClass}.}

\item{...}{Additional graphical parameters passed to the plotting functions.}
}
\value{
No return value. Called for its side effect (producing plots).
}
\description{
Visualizes each variable in a hybrid data object (\code{hdClass}) using an appropriate plot style:
\itemize{
\item Functional data are shown using lines (\code{matplot(..., type = "l")}).
\item Regular data are shown using solid dots (\code{matplot(..., type = "p", pch = 16)}).
\item Image data are plotted using \code{image()} if they originated from matrices.
}
}
\examples{
fd_obj <- fdClass(matrix(rnorm(100), 10, 10))
rd_obj <- rdClass(matrix(rnorm(100), 10, 10))
hd_obj <- hdClass(list(fd_obj, rd_obj))
plot(hd_obj)

}
