% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgMethod.R
\name{plot.imgClass}
\alias{plot.imgClass}
\title{Plot Method for imgClass Objects}
\usage{
\method{plot}{imgClass}(x)
}
\arguments{
\item{x}{An object of class \code{imgClass}, created using \code{imgClass()}.}

\item{col}{Color palette to use for image rendering (default: grayscale).}
}
\value{
No return value. Called for its side effect (plotting).
}
\description{
Visualizes a list of image matrices stored in an \code{imgClass} object.
Each image is shown one at a time. The user is prompted to press Enter or
click on the graphics window to advance to the next image.
}
\examples{
img_list <- list(matrix(rnorm(100), 10, 10),
                 matrix(runif(100), 10, 10))
img_obj <- imgClass(img_list)
plot(img_obj)

}
