% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMPCAMethod.R
\name{plot_cv_u}
\alias{plot_cv_u}
\title{Plot CV Error with 1-SE Rule for u}
\usage{
plot_cv_u(ReMPCA_obj)
}
\arguments{
\item{ReMPCA_obj}{Output list from the ReMPCA routine.}
}
\value{
Side-by-side base R plots for each component showing the CV error and 1-SE threshold.
}
\description{
For each PC component, plots CV error vs. sparsity tuning parameter (\code{gamma})
and highlights the optimal value selected via the 1-SE rule.
}
