% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMPCAMethod.R
\name{plot_gcv_u}
\alias{plot_gcv_u}
\title{Plot GCV Scores for U}
\usage{
plot_gcv_u(ReMPCA_obj, ...)
}
\arguments{
\item{ReMPCA_obj}{Output list from ReMPCA.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
This function visualizes the GCV scores for the left singular vectors (U)
across different regularization values.
}
