% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMPCAMethod.R
\name{plot_gcv_v}
\alias{plot_gcv_v}
\title{Plot GCV Results for Functional Loadings (\code{v}) in ReMPCA}
\usage{
plot_gcv_v(ReMPCA_obj, ...)
}
\arguments{
\item{ReMPCA_obj}{A list-like ReMPCA object with GCVResultsV and OptimalAlphaV.}

\item{...}{Additional plotting options.}
}
\description{
This function visualizes the Generalized Cross-Validation (GCV) curves or surfaces
for the functional loading vectors (\code{v}) across components in a ReMPCA object.
}
