% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdClass.R
\name{rdClass}
\alias{rdClass}
\title{Regular Data Class}
\usage{
rdClass(data, Sparsity_parameter = 0)
}
\arguments{
\item{data}{A matrix representing the data, with rows indicating observations and columns representing variables.}

\item{Sparsity_parameter}{\itemize{
\item A fixed number representing the level of sparsity for columns, or a vector of numerical values that will undergo cross-validation (CV) to determine the optimal value.
\item For no sparsity, set it to 0.
\item If \code{NULL}, the sparsity parameter will be tuned automatically.
}}
}
\description{
The \code{rdClass} class denotes an element of regular data.
\itemize{
\item Data objects are constructed with observations in rows and variables in columns.
\item The smoothing parameter is automatically set to zero (no smoothing).
\item Users can assign sparsity parameters only!
\item No grid points are involved in \code{rdClass} objects.
}
}
\examples{
# Example for Regular Data (rd)
rd_data <- matrix(rnorm(100), nrow = 10, ncol = 10)  # 10 rows, 10 columns
rd_object <- rdClass(data = rd_data,
                     Sparsity_parameter = 3)  # Custom sparsity parameter

# Display the created rd object
print(rd_object)
print(attr(rd_object, "Sparsity_parameter"))  # Display sparsity parameter
}
