% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{[.hdClass}
\alias{[.hdClass}
\title{Subset Method for hdClass}
\usage{
\method{[}{hdClass}(x, i = NULL, j = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{hdClass}.}

\item{i}{Row indices (observations). If \code{NULL}, all rows are included.}

\item{j}{Column indices (variables). If \code{NULL}, all variables are included.}

\item{...}{Not used.}
}
\value{
A new \code{hdClass} object with subsetted data and inherited attributes.
}
\description{
Enables subsetting of an \code{hdClass} object by rows (observations) and columns (variables).
}
